First README

+++ MainWindow.py
@@ -33,12 +33,15 @@ class MainWindow(QMainWindow):
     def close_app(self):
         folder = 'equations'
         for the_file in os.listdir(folder):
-            file_path = os.path.join(folder, the_file)
-            try:
-                if os.path.isfile(file_path):
-                    os.unlink(file_path)
-                #elif os.path.isdir(file_path): shutil.rmtree(file_path)
-            except Exception as e:
-                print(e)
+            
+            # Delete all files except about.txt
+            if(str(the_file) != "about.txt"):
+                file_path = os.path.join(folder, the_file)
+                try:
+                    if os.path.isfile(file_path):
+                        os.unlink(file_path)
+                    #elif os.path.isdir(file_path): shutil.rmtree(file_path)
+                except Exception as e:
+                    print(e)
                 
         self.close()
